save data on local database

+++ app/src/main/java/br/com/raynerweb/pokemon/repository/local/entity/PokemonEntity.kt
@@ -6,7 +6,7 @@ import androidx.room.PrimaryKey
 @Entity(tableName = "pokemon")
 data class PokemonEntity(
     @PrimaryKey(autoGenerate = true)
-    val pokemonId: Int = 0,
+    val pokemonId: Long = 0,
     val image: String,
     val name: String,
 )
